chore: Rename services files and change file ext.
